pyproject.toml
[tool.poetry]
name = "repl_nix_EGFiuza-Careers"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
SQLAlchemy = "^2.0.12"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
